DEV: Moved vmtk tcl functions to vmtk utils
