+++ Code/Source/Mesh/VMTKUtils/cv_vmtk_utils.cxx
@@ -57,6 +57,7 @@
 #include "vtkDataSetSurfaceFilter.h"
 #include "vtkGeometryFilter.h"
 #include "vtkCellArray.h"
+#include "vtkSVFillHolesFilterWithIds.h"
 
 #include "vtkvmtkPolyDataSurfaceRemeshing.h"
 #include "vtkvmtkPolyDataSizingFunction.h"
@@ -66,6 +67,599 @@
 #include "vtkvmtkAppendFilter.h"
 #include "vtkvmtkBoundaryLayerGenerator.h"
 #include "cv_polydatasolid_utils.h"
+#include "vtkvmtkPolyDataDistanceToCenterlines.h"
+#include "vtkvmtkPolyDataCenterlines.h"
+#include "vtkvmtkCenterlineBranchExtractor.h"
+#include "vtkvmtkCapPolyData.h"
+#include "vtkvmtkSimpleCapPolyData.h"
+#include "vtkvmtkPolyDataCenterlineGroupsClipper.h"
+#include "vtkvmtkMergeCenterlines.h"
+
+#define vtkNew(type,name) \
+  vtkSmartPointer<type> name = vtkSmartPointer<type>::New()
+
+/* -------------- */
+/* sys_geom_centerlines */
+/* -------------- */
+
+/** @author Adam Updegrove
+ *  @author updega2@gmail.com
+ *  @author UC Berkeley
+ *  @author shaddenlab.berkeley.edu
+ *
+ *  @brief Function to extract centerlines from a vtkPolyData surface
+ *  @brief VTMK is called to do this. Each cap on PolyData has an
+ *  @brief id and then each id is set as either inlet or outlet.
+ *  @param *sources list of source cap ids
+ *  @param nsources number of source cap ids
+ *  @param *targets list of target cap ids
+ *  @param ntargets number of target cap ids
+ *  @param **lines returned center lines as vtkPolyData
+ *  @param ** voronoi returned voronoi diagram as vtkPolyData
+ *  @return SV_OK if the VTMK function executes properly
+ */
+
+int sys_geom_centerlines( cvPolyData *polydata,int *sources,int nsources,
+		int *targets,int ntargets,
+		cvPolyData **lines, cvPolyData **voronoi)
+{
+  vtkPolyData *geom = polydata->GetVtkPolyData();
+  cvPolyData *result1 = NULL;
+  cvPolyData *result2 = NULL;
+  *lines = NULL;
+  *voronoi = NULL;
+
+//  vtkSmartPointer<vtkvmtkPolyDataCenterlines> centerLiner =
+//    vtkSmartPointer<vtkvmtkPolyDataCenterlines>::New();
+  vtkSmartPointer<vtkIdList> capInletIds = vtkSmartPointer<vtkIdList>::New();
+  vtkSmartPointer<vtkIdList> capOutletIds = vtkSmartPointer<vtkIdList>::New();
+  int pointId;
+
+  fprintf(stderr,"NumSources %d\n",nsources);
+  fprintf(stderr,"NumTargets %d\n",ntargets);
+  for (pointId = 0;pointId<nsources;pointId++)
+  {
+    capInletIds->InsertNextId(*sources+pointId);
+  }
+  for (pointId = 0;pointId<ntargets;pointId++)
+  {
+    capOutletIds->InsertNextId(*targets+pointId);
+  }
+
+  vtkNew(vtkvmtkPolyDataCenterlines,centerLiner);
+  try {
+    std::cout<<"Getting Center Lines..."<<endl;
+    centerLiner->SetInputData(geom);
+    centerLiner->SetSourceSeedIds(capInletIds);
+    centerLiner->SetTargetSeedIds(capOutletIds);
+    centerLiner->SetRadiusArrayName("MaximumInscribedSphereRadius");
+    centerLiner->SetCostFunction("1/R");
+    centerLiner->SetFlipNormals(0);
+    centerLiner->SetAppendEndPointsToCenterlines(1);
+    centerLiner->SetSimplifyVoronoi(0);
+    centerLiner->SetCenterlineResampling(0);
+    centerLiner->SetResamplingStepLength(1);
+    centerLiner->Update();
+
+    result1 = new cvPolyData( centerLiner->GetOutput() );
+    *lines = result1;
+    result2 = new cvPolyData( centerLiner->GetVoronoiDiagram() );
+    *voronoi = result2;
+  }
+  catch (...) {
+    fprintf(stderr,"ERROR in centerline operation.\n");
+    fflush(stderr);
+    return SV_ERROR;
+  }
+
+  return SV_OK;
+}
+
+/* -------------- */
+/* sys_geom_mergecenterlines */
+/* -------------- */
+
+/** @author Adam Updegrove
+ *  @author updega2@gmail.com
+ *  @author UC Berkeley
+ *  @author shaddenlab.berkeley.edu
+ *
+ *  @brief Function to merge centerlines from a vtkPolyData surface
+ *  @brief Must be called after separation of centerlines into groups
+ *  @brief VTMK is called to do this.
+ *  @param *sources list of source cap ids
+ *  @param nsources number of source cap ids
+ *  @param *targets list of target cap ids
+ *  @param ntargets number of target cap ids
+ *  @param **lines returned center lines as vtkPolyData
+ *  @param ** voronoi returned voronoi diagram as vtkPolyData
+ *  @return SV_OK if the VTMK function executes properly
+ */
+
+int sys_geom_mergecenterlines( cvPolyData *lines, int mergeblanked,
+		cvPolyData **merged)
+{
+  vtkPolyData *geom = lines->GetVtkPolyData();
+  cvPolyData *result1 = NULL;
+  *merged = NULL;
+
+  vtkNew(vtkvmtkMergeCenterlines, merger);
+  try {
+    std::cout<<"Merging Sections..."<<endl;
+    merger->SetInputData(geom);
+    merger->SetBlankingArrayName("Blanking");
+    merger->SetRadiusArrayName("MaximumInscribedSphereRadius");
+    merger->SetGroupIdsArrayName("GroupIds");
+    merger->SetCenterlineIdsArrayName("CenterlineIds");
+    merger->SetTractIdsArrayName("TractIds");
+    merger->SetMergeBlanked(mergeblanked);
+    merger->Update();
+
+    result1 = new cvPolyData( merger->GetOutput() );
+    *merged = result1;
+  }
+  catch (...) {
+    fprintf(stderr,"ERROR in centerline merging.\n");
+    fflush(stderr);
+    return SV_ERROR;
+  }
+
+  return SV_OK;
+
+  return SV_OK;
+}
+
+/* -------------- */
+/* sys_geom_separatecenterlines */
+/* -------------- */
+
+/** @author Adam Updegrove
+ *  @author updega2@gmail.com
+ *  @author UC Berkeley
+ *  @author shaddenlab.berkeley.edu
+ *
+ *  @brief Function to separate centerlines based on bifurcations of the geometry
+ *  @param *lines from centerline extraction, the centerlines to be used
+ *  @note  The output lines are not physically different than the input. The
+ *  @note  simply applies cell data to the lines that describe the bifurcation regions.
+ *  @note  The cell array names are "Blanking","GroupIds","CenterlineIds",and "TractIds"
+ *  @param **separate vtkPolyData with cell arrays discribing branchin attached
+ *  @return SV_OK if the VTMK function executes properly
+ */
+
+int sys_geom_separatecenterlines( cvPolyData *lines,
+		cvPolyData **separate)
+{
+  vtkPolyData *geom = lines->GetVtkPolyData();
+  cvPolyData *result1 = NULL;
+  *separate = NULL;
+
+  vtkNew(vtkvmtkCenterlineBranchExtractor,brancher);
+  try {
+    std::cout<<"Grouping Sections..."<<endl;
+    brancher->SetInputData(geom);
+    brancher->SetBlankingArrayName("Blanking");
+    brancher->SetRadiusArrayName("MaximumInscribedSphereRadius");
+    brancher->SetGroupIdsArrayName("GroupIds");
+    brancher->SetCenterlineIdsArrayName("CenterlineIds");
+    brancher->SetTractIdsArrayName("TractIds");
+    brancher->Update();
+
+    result1 = new cvPolyData( brancher->GetOutput() );
+    *separate = result1;
+  }
+  catch (...) {
+    fprintf(stderr,"ERROR in centerline separation.\n");
+    fflush(stderr);
+    return SV_ERROR;
+  }
+
+  return SV_OK;
+}
+
+/* -------------- */
+/* sys_geom_grouppolydata */
+/* -------------- */
+
+/** @author Adam Updegrove
+ *  @author updega2@gmail.com
+ *  @author UC Berkeley
+ *  @author shaddenlab.berkeley.edu
+ *
+ *  @brief Function to calculate distance to surface from points on line
+ *  @param *polydata The polydata to find distance form lines on
+ *  @param *lines from centerline extraction, the centerlines to be used
+ *  @param **distance vtkPolyData with distance functiona attached returned
+ *  @return SV_OK if the VTMK function executes properly
+ */
+
+int sys_geom_grouppolydata( cvPolyData *polydata,cvPolyData *lines,
+		cvPolyData **grouped)
+{
+  vtkPolyData *geom = polydata->GetVtkPolyData();
+  vtkPolyData *centerlines = lines->GetVtkPolyData();
+  cvPolyData *result = NULL;
+  *grouped = NULL;
+
+  vtkNew(vtkvmtkPolyDataCenterlineGroupsClipper,grouper);
+  try {
+    std::cout<<"Branching PolyData..."<<endl;
+    grouper->SetInputData(geom);
+    grouper->SetCenterlines(centerlines);
+    grouper->SetCenterlineGroupIdsArrayName("GroupIds");
+    grouper->SetGroupIdsArrayName("GroupIds");
+    grouper->SetCenterlineRadiusArrayName("MaximumInscribedSphereRadius");
+    grouper->SetBlankingArrayName("Blanking");
+    grouper->SetCutoffRadiusFactor(1e16);
+    grouper->SetClipValue(0.0);
+    grouper->SetUseRadiusInformation(1);
+    grouper->ClipAllCenterlineGroupIdsOn();
+    grouper->Update();
+
+    result = new cvPolyData( grouper->GetOutput());
+    *grouped = result;
+  }
+  catch (...) {
+    fprintf(stderr,"ERROR in centerline operation.\n");
+    fflush(stderr);
+    return SV_ERROR;
+  }
+  return SV_OK;
+}
+
+/* -------------- */
+/* sys_geom_distancetocenterlines */
+/* -------------- */
+
+/** @author Adam Updegrove
+ *  @author updega2@gmail.com
+ *  @author UC Berkeley
+ *  @author shaddenlab.berkeley.edu
+ *
+ *  @brief Function to calculate distance to surface from points on line
+ *  @param *polydata The polydata to find distance form lines on
+ *  @param *lines from centerline extraction, the centerlines to be used
+ *  @param **distance vtkPolyData with distance functiona attached returned
+ *  @return SV_OK if the VTMK function executes properly
+ */
+
+int sys_geom_distancetocenterlines( cvPolyData *polydata,cvPolyData *lines,
+		cvPolyData **distance)
+{
+  vtkPolyData *geom = polydata->GetVtkPolyData();
+  vtkPolyData *centerlines = lines->GetVtkPolyData();
+  cvPolyData *result = NULL;
+  *distance = NULL;
+
+  vtkNew(vtkvmtkPolyDataDistanceToCenterlines,distancer);
+  try {
+    std::cout<<"Getting Distance to Center Lines..."<<endl;
+    distancer->SetInputData(geom);
+    distancer->SetCenterlines(centerlines);
+    distancer->SetUseRadiusInformation(1);
+    distancer->SetEvaluateTubeFunction(0);
+    distancer->SetEvaluateCenterlineRadius(0);
+    distancer->SetProjectPointArrays(0);
+    distancer->SetDistanceToCenterlinesArrayName("DistanceToCenterlines");
+    distancer->SetCenterlineRadiusArrayName("MaximumInscribedSphereRadius");
+    distancer->Update();
+
+    result = new cvPolyData( distancer->GetOutput());
+    *distance = result;
+  }
+  catch (...) {
+    fprintf(stderr,"ERROR in centerline operation.\n");
+    fflush(stderr);
+    return SV_ERROR;
+  }
+  return SV_OK;
+}
+
+/* -------------- */
+/* sys_geom_cap */
+/* -------------- */
+
+/** @author Adam Updegrove
+ *  @author updega2@gmail.com
+ *  @author UC Berkeley
+ *  @author shaddenlab.berkeley.edu
+ *
+ *  @brief Function that calls VMTK method to cap which assigns ids t caps
+ *  @param *polydata The polydata to cap
+ *  @param *cappedpolydata the capped polydata is returned in this
+ *  @param *numcenterids Number of caps or ids for the center of caps
+ *  @param **centerids If type is 1, a cap with a centerpoint is added.
+ *  Then this centerpoint id is return in this object
+ *  @param type This determines whether to cap regularly or cap with a point
+ *  in the center. 0 - regular, 1 - point in center
+ *  @return SV_OK if the VTMK function executes properly
+ */
+
+int sys_geom_cap( cvPolyData *polydata,cvPolyData **cappedpolydata,int *numcenterids,int **centerids,int type)
+{
+  vtkPolyData *geom = polydata->GetVtkPolyData();
+  cvPolyData *result = NULL;
+  *cappedpolydata = NULL;
+  vtkSmartPointer<vtkIdList> capCenterIds = vtkSmartPointer<vtkIdList>::New();
+  vtkSmartPointer<vtkTriangleFilter> triangulate =
+	  vtkSmartPointer<vtkTriangleFilter>::New();
+  int numids;
+  int *allids;
+  int i;
+
+  try {
+    if (type ==0)
+    {
+      vtkSmartPointer<vtkvmtkSimpleCapPolyData> capper =
+	      vtkSmartPointer<vtkvmtkSimpleCapPolyData>::New();
+      capper->SetInputData(geom);
+      capper->SetCellEntityIdsArrayName("CenterlineCapID");
+      capper->SetCellEntityIdOffset(1);
+      capper->Update();
+      triangulate->SetInputData(capper->GetOutput());
+      triangulate->Update();
+
+      result = new cvPolyData( triangulate->GetOutput() );
+      *cappedpolydata = result;
+      capCenterIds->InsertNextId(0);
+    }
+    else if (type == 1)
+    {
+      vtkSmartPointer<vtkvmtkCapPolyData> capper =
+	      vtkSmartPointer<vtkvmtkCapPolyData>::New();
+      capper->SetInputData(geom);
+      capper->SetDisplacement(0);
+      capper->SetInPlaneDisplacement(0);
+      capper->Update();
+      triangulate->SetInputData(capper->GetOutput());
+      triangulate->Update();
+
+      result = new cvPolyData( triangulate->GetOutput() );
+      *cappedpolydata = result;
+      capCenterIds->DeepCopy(capper->GetCapCenterIds());
+
+    }
+
+  }
+  catch (...) {
+    fprintf(stderr,"ERROR in capping operation.\n");
+    fflush(stderr);
+    return SV_ERROR;
+  }
+
+  numids = capCenterIds->GetNumberOfIds();
+  allids = new int[numids];
+
+  for (i=0;i<numids;i++)
+  {
+    allids[i] = capCenterIds->GetId(i);
+  }
+  *numcenterids = numids;
+  *centerids = allids;
+
+  return SV_OK;
+}
+
+/* -------------- */
+/* sys_geom_cap_with_ids */
+/* -------------- */
+
+/** @author Adam Updegrove
+ *  @author updega2@gmail.com
+ *  @author UC Berkeley
+ *  @author shaddenlab.berkeley.edu
+ *
+ *  @brief Function that calls customized vtk filter to cap with ids starting
+ *  from the lowest number in the current "ModelFaceID" array
+ *  @param *polydata The polydata to cap
+ *  @param fillid the id number to use for capping. Depends on filltype
+ *  @param filledholes the number of holes filled in process
+ *  @param filltype How to assign ids to caps on polydata. 0 - start from 1
+ *  and each cap gets a new id. 1 - give every cap the same id (fillid).
+ *  2 - start from fillid and give each new cap a value increasing from this
+ *  @param type This determines whether to cap regularly or cap with a point
+ *  in the center. 0 - regular, 1 - point in center
+ *  @return SV_OK if the VTMK function executes properly
+ */
+
+int sys_geom_cap_with_ids( cvPolyData *polydata,cvPolyData **cappedpolydata,
+    int fillId,int filledholes,int filltype)
+{
+
+  vtkPolyData *geom = polydata->GetVtkPolyData();
+  cvPolyData *result = NULL;
+  *cappedpolydata = NULL;
+
+  try {
+
+    std::cout<<"Capping Surface..."<<endl;
+    vtkSmartPointer<vtkSVFillHolesFilterWithIds> capper =
+	    vtkSmartPointer<vtkSVFillHolesFilterWithIds>::New();
+    capper->SetInputData(geom);
+    capper->SetFillId(fillId);
+    //Fill type, 0 for number of holes filled, 1 for a fillid, and 2 for
+    //increasing number id starting at given fillid number+1
+    capper->SetFillType(filltype);
+    capper->SetHoleSize(capper->GetHoleSizeMaxValue());
+    capper->Update();
+
+    filledholes = capper->GetNumberOfHolesFilled();
+    vtkNew(vtkPolyData,capout);
+    capout->DeepCopy(capper->GetOutput());
+
+    if (filltype ==2)
+    {
+      vtkNew(vtkIntArray, currentCapArray);
+      vtkNew(vtkIntArray, currentFaceArray);
+      vtkNew(vtkIntArray, newFaceArray);
+      if (VtkUtils_PDCheckArrayName(capout,1,"CapID") != SV_OK)
+      {
+	fprintf(stderr,"CapID Array is not on the surface\n");
+	return SV_ERROR;
+      }
+      if (VtkUtils_PDCheckArrayName(geom,1,"ModelFaceID") != SV_OK)
+      {
+	fprintf(stderr,"ModelFaceID Array is not on the surface\n");
+	return SV_ERROR;
+      }
+      currentCapArray = vtkIntArray::SafeDownCast(capout->GetCellData()->
+	  GetArray("CapID"));
+      currentFaceArray = vtkIntArray::SafeDownCast(geom->GetCellData()->
+	  GetArray("ModelFaceID"));
+      for (int i=0;i<capout->GetNumberOfCells();i++)
+      {
+	if (currentCapArray->GetValue(i) == -1)
+	  newFaceArray->InsertValue(i,currentFaceArray->GetValue(i));
+	else
+	  newFaceArray->InsertValue(i,currentCapArray->GetValue(i));
+      }
+      newFaceArray->SetName("ModelFaceID");
+      capout->GetCellData()->AddArray(newFaceArray);
+      capout->GetCellData()->SetActiveScalars("ModelFaceID");
+    }
+    result = new cvPolyData(capout);
+    *cappedpolydata = result;
+  }
+  catch (...) {
+    fprintf(stderr,"ERROR in capping operation.\n");
+    fflush(stderr);
+    return SV_ERROR;
+  }
+
+  return SV_OK;
+}
+
+/* -------------- */
+/* sys_geom_mapandcorrectids */
+/* -------------- */
+
+/** @author Adam Updegrove
+ *  @author updega2@gmail.com
+ *  @author UC Berkeley
+ *  @author shaddenlab.berkeley.edu
+ *
+ *  @brief Function to map and correct ids after a surface has been thresholded
+ *  then capped.
+ *  @param *polydata The polydata map and correct ids on
+ *  @param *originalpd The original polydata with correct ids
+ *  @param *newpd The new polydata to correct the ids on
+ *  @param **polydata The returned polydata with everything all set
+ *  @param *originalarray original id array name on the surface to correct
+ *  @param *newarray new id array name to give to the corrected ids
+ *  @return SV_OK if the VTMK function executes properly
+ */
+
+int sys_geom_mapandcorrectids( cvPolyData *originalpd, cvPolyData *newpd, cvPolyData **polydata,char *originalarray, char *newarray)
+{
+  vtkPolyData *originalgeom = originalpd->GetVtkPolyData();
+  vtkPolyData *newgeom = newpd->GetVtkPolyData();
+  cvPolyData *result = NULL;
+  *polydata = NULL;
+
+  int i,j,k;
+  int subId;
+  int count;
+  vtkIdType npts;
+  vtkIdType *pts;
+  double distance;
+  double closestPt[3];
+  double minmax[2];
+  double centroid[3];
+  long range;
+  vtkIdType closestCell;
+  vtkIdType cellId;
+  vtkIdType currentValue;
+  vtkPolyData *newcopy = vtkPolyData::New();
+  vtkSmartPointer<vtkCellLocator> locator =
+    vtkSmartPointer<vtkCellLocator>::New();
+  vtkSmartPointer<vtkPointLocator> pointLocator =
+    vtkSmartPointer<vtkPointLocator>::New();
+  vtkSmartPointer<vtkGenericCell> genericCell =
+    vtkSmartPointer<vtkGenericCell>::New();
+  vtkSmartPointer<vtkLongArray> currentRegionsLong =
+    vtkSmartPointer<vtkLongArray>::New();
+  vtkSmartPointer<vtkIntArray> currentRegionsInt =
+    vtkSmartPointer<vtkIntArray>::New();
+  vtkSmartPointer<vtkIntArray> realRegions =
+    vtkSmartPointer<vtkIntArray>::New();
+  vtkSmartPointer<vtkIdList> closestCells =
+    vtkSmartPointer<vtkIdList>::New();
+
+  newcopy->DeepCopy(newgeom);
+  newcopy->BuildLinks();
+  originalgeom->BuildLinks();
+  locator->SetDataSet(originalgeom);
+  locator->BuildLocator();
+  pointLocator->SetDataSet(originalgeom);
+  pointLocator->BuildLocator();
+
+  currentRegionsLong = static_cast<vtkLongArray*>(newcopy->GetCellData()->GetScalars(newarray));
+  currentRegionsLong->GetRange(minmax,0);
+
+  realRegions = static_cast<vtkIntArray*>(originalgeom->GetCellData()->GetScalars(originalarray));
+//  realRegions = vtkIntArray::SafeDownCast(originalgeom->GetCellData()->GetScalars(originalarray));
+
+  range = minmax[1]-minmax[0];
+  long *mapper;
+  mapper = new long[1+range];
+
+  for (i=0;i<range+1;i++)
+  {
+    mapper[i] = -1;
+  }
+
+  for (cellId=0;cellId<newcopy->GetNumberOfCells();cellId++)
+  {
+    currentValue = currentRegionsLong->GetValue(cellId);
+
+    if (mapper[currentValue-1] == -1)
+    {
+      fprintf(stderr,"Getting Value: %d\n",currentValue-1);
+      newcopy->GetCellPoints(cellId,npts,pts);
+
+      vtkSmartPointer<vtkPoints> polyPts = vtkSmartPointer<vtkPoints>::New();
+      vtkSmartPointer<vtkIdTypeArray> polyPtIds = vtkSmartPointer<vtkIdTypeArray>::New();
+      for (i=0;i<npts;i++)
+      {
+	polyPtIds->InsertValue(i,i);
+	polyPts->InsertNextPoint(newcopy->GetPoint(pts[i]));
+      }
+
+      vtkPolygon::ComputeCentroid(polyPtIds,polyPts,centroid);
+
+      locator->FindClosestPoint(centroid,closestPt,genericCell,closestCell,
+	  subId,distance);
+
+      fprintf(stderr,"the value: %d\n",realRegions->GetValue(closestCell));
+      mapper[currentValue-1] = realRegions->GetValue(closestCell);
+    }
+  }
+
+  for (i=0;i<range+1;i++)
+  {
+    fprintf(stderr,"Want to see mapper vals: %d is %d\n",i,mapper[i]);
+  }
+
+  //Set original region values
+  for (cellId=0;cellId<newcopy->GetNumberOfCells();cellId++)
+  {
+    currentValue = static_cast<int>(currentRegionsLong->GetValue(cellId));
+    currentRegionsInt->InsertValue(cellId,mapper[currentValue-1]);
+  }
+
+  newcopy->GetCellData()->RemoveArray(newarray);
+
+  currentRegionsInt->SetName(originalarray);
+  newcopy->GetCellData()->AddArray(currentRegionsInt);
+  newcopy->GetCellData()->SetActiveScalars(originalarray);
+
+  result = new cvPolyData( newcopy );
+  *polydata = result;
+
+  delete [] mapper;
+  return SV_OK;
+}
+
 // --------------------
 //  VMTKUtils_SurfaceRemeshing
 // --------------------
